+++ webknossos/webknossos/client/_generated/models/__init__.py
@@ -134,6 +134,19 @@ from .dataset_info_response_200_data_source_data_layers_item_default_view_config
 )
 from .dataset_info_response_200_data_source_id import DatasetInfoResponse200DataSourceId
 from .dataset_info_response_200_data_store import DatasetInfoResponse200DataStore
+from .dataset_list_response_200_item import DatasetListResponse200Item
+from .dataset_list_response_200_item_allowed_teams_item import (
+    DatasetListResponse200ItemAllowedTeamsItem,
+)
+from .dataset_list_response_200_item_data_source import (
+    DatasetListResponse200ItemDataSource,
+)
+from .dataset_list_response_200_item_data_source_id import (
+    DatasetListResponse200ItemDataSourceId,
+)
+from .dataset_list_response_200_item_data_store import (
+    DatasetListResponse200ItemDataStore,
+)
 from .dataset_reserve_upload_json_body import DatasetReserveUploadJsonBody
 from .dataset_sharing_token_response_200 import DatasetSharingTokenResponse200
 from .dataset_update_json_body import DatasetUpdateJsonBody
@@ -180,6 +193,7 @@ from .task_infos_by_project_id_response_200_item_type_settings import (
 from .task_infos_by_project_id_response_200_item_type_settings_resolution_restrictions import (
     TaskInfosByProjectIdResponse200ItemTypeSettingsResolutionRestrictions,
 )
+from .team_list_response_200_item import TeamListResponse200Item
 from .user_info_by_id_response_200 import UserInfoByIdResponse200
 from .user_info_by_id_response_200_experiences import UserInfoByIdResponse200Experiences
 from .user_info_by_id_response_200_novel_user_experience_infos import (